Show ruby warnings even if syntax OK
On Ruby 2.1.2 the "Syntax OK" message is displayed even if there are warnings (such as unused variables) in the file - so potentially useful warnings would be hidden. This change modifies the `match_ok` regex to use multiline matching and a negative lookbehind assertion to verify that there is not a newline before the "Syntax OK" message.